+++ include/kyosu/types/impl/bessel.hpp
@@ -44,6 +44,7 @@
 #include <kyosu/types/impl/detail/bessel_y.hpp>
 #include <kyosu/types/impl/detail/bessel_h.hpp>
 #include <kyosu/types/impl/detail/bessel_k.hpp>
+#include <kyosu/types/impl/detail/bessel_jy.hpp>
 // bessels of floating (real) order
 #include <kyosu/types/impl/detail/bessel_jr.hpp>
 #include <kyosu/types/impl/detail/bessel_yr.hpp>
